move worker connection impls into worker_connection.dart, updated comments to be more relevant
